Added BASH script for basic Linux virus.
This virus program will infect the first Linux ELF executable that it finds in the current directory (in any order). Note that this will only infect ELF executables and not executable scripts or similar. This virus will also reproduce by itself, i.e., whenever an infected program is executed, the virus should find another file and should infect it. Furthermore, files that have already been infected would not be infected again.

The "payload" of this virus is a simple message for academic purposes. Every time an application that is infected is started, it would print "Hello! I am a simple virus!\n" to <stdout> before or after it completes its task.

Of course, needless to say, the original functionality of the host application (i.e., the program that got infected) would not change. All command-line arguments should work. For the sake of simplicity, I have ignored environment variables.